v5
This is a Beta stage of the version 5 release of Kardna. Along side this we will be releasing Oasis OS. These to website are the future for our network of applications and we hope to develop more in the future. We also in the future plan for a API for other people to use the assets of the sites and apps we make and implement them into their own.